fix(shell): .profile get /bin/sh .bashrc gets rest
Fix so that .profile only has /bin/sh commands so will work in Ubuntu.
Also fixes so that .bashrc gets completion commands that are for bash
These rerun with each .bashrc. The completions maybe should go into
.bash_profile so they are not executed over and over.

There are PATHS which work fine, but the sourced completions have
fucntions and other things such as in yq completion, you would need to
examine each to know if it could bo into .bash_profile instead of
.bashrc