fix: Ubuntu for non-rich installs
These fixes work with the demo account and fix things that such as

1Password is now installed from their private repo. The snap version is
sandboxed so does not work with browser extension and does not have the
op cli

Fixes so that gcc is actually used on the MacOS, but not on Ubuntu since
asdf install python does not work just with homebrew gcc, it works with
the apt packages

install-conda works correctly for linux but the source of .bashrc does
not work correctly but we do not need this since we do not need to work
outside of the conda environment. So just comment out conda deactivate

install-git-tools was only using the -p entry for email and  ignoring
the -e, so got rid of -p entirely so the .gitconfig gets the right email
address when called by installs.h

install-gnu.sh now installs gcc and has a FORCE so that you can link to
it. This seems to break the Ubuntu toolchain for asdf so it is not the
default.

install-google-drive.sh now installs the fuse drive so that on Ubuntu
you can access the Google Drive. Good for small things and things that
are not to be checked in.

Adding more packages to install like golang and ripgrep

install-conda had a typo in the config_add
powerline installation has warning and moved plugins to .zshrc
ruby has a typo in the config addition