fix: vimrc pipenv quiet, ssh-config install hang
One of the libraries in install-ssh-config is hanging so bisecting to
see which one it is

vimrc install changes to run a PIPENV_QUITE in case there is a .env file
that sets Python pipenv variables. Note that this overlaps with asdf
direnv which uses .envrc for the same purpose, so you don't need .env
most of the time, use .envrc instead.