fix: install asdf updates plugins
Really frustrating error with asdf direnv setup --shell bash it kept
saying cannot update shell submit a PR, but what was happening is that
on an older MacBook that didn't have the latest direnv plugin it did not
recognize the --shell variable at all and didn't pass it along.

So no install-asdf.sh if it detects a plugin will automatically update
it.