feat: exif-copy, gawk path, add kfp
exif-copy for use with darktable because .jpg exports have exif but
.avif do not so this copies the exif information from .jpg into .avif.
It does require two exports, but you get the data you need. Issue filed
with darktable folks. The same is true of .jp2 as well as .

brew installed gawk needs an explicit path to access the

Added kfp the Kubeflow Pipeline CLI to kubernetes installation.