feat: Unity installation, brew link added
Unity first cut at installation this fails because Unity Hub does not
have a license file but it does work with a manual installation so the
fix is to install manually the brew reinstall to keep it up to date.
Also added code for a direct package installation.

In debugging checked brew doctor and found several cases where keg only
installations were not linked, there was just a copy of the file in
homebrew bin. So fixed that with manual links.