feat: Yubikey for linux (partial) and mac and appimage
Add appimage launcher and uses appimage_install moving that code out of
install-qgc to be more generally used since appimages and snaps are more
linux version independent than apt repositories.

If you have Yubikey this will enable it for the Mac so you will need the
key and a pincode to login. This is not that much use on modern Macs
with Touch authentication but could be useful for older macs. It adds
the Yubikey Manager to change your pins, the Yubikey Authenticator if
you want to store your 2FA keys there.

Note that you really want two keys at least and probably more so if you
lose them you won't be locked out. So much less useful than 1Password
but more secure as the physical key and pin are needed. Tested with
Yubikey 4C, but using the one with fingerprint identification is even
better.

Also has rough untested support for Linux to install the manager and the
authenticator but does no implement the /etc/pam.d and pam.so which
would require using the keys. Macs just allow this as an additional
factor rather than requiring it. Also does not secure the u2f_keys which
are the public keys nor does it put them into /etc as it should if you
are serious.