fix: aws and ssh configuration fixes
Adding aws sso configuration which generates a temporary token for
applications that are local. Bug fix as the sso code isn't running
regularly in install-aws-config.sh

Documentation to note that 1Password is the default for ssh keys and
that sso is the default for aws tokens with the experiment with
1password authentication in process.

Using sso with temporary credentials now works. Note that the 1Password
one is not quite working yet.