feat: 11template and chezmoi
Adding chezmoi to manage dotfiles. If it works well then can drop my own
dotfiles. Main complexity is that this means I need multiple dotfile
repos (not that big a deal). And it works with dotfiles-stow.sh which is
nice

Also adding a template as 1password which comes up first is too
complicated.

This is called 11template.sh so is alphabetically before 1Password